+++ CHANGELOG.md
@@ -1,6 +1,22 @@
 Changelog
 =========
 
+## v6.21.0 (23 June 2021)
+
+### Enhancements
+
+* Allow a `Method` or any object responding to `#call` to be used as an `on_error` callback or middleware
+  | [#662](https://github.com/bugsnag/bugsnag-ruby/pull/662)
+  | [odlp](https://github.com/odlp)
+
+### Fixes
+
+* Deliver when an error is raised in the block argument to `notify`
+  | [#660](https://github.com/bugsnag/bugsnag-ruby/pull/660)
+  | [aki77](https://github.com/aki77)
+* Fix potential `NoMethodError` in `Bugsnag::Railtie` when using `require: false` in a Gemfile
+  | [#666](https://github.com/bugsnag/bugsnag-ruby/pull/666)
+
 ## v6.20.0 (29 March 2021)
 
 ### Enhancements